Move more code into `Features`
* Move `KeyedCipherStream` to the `Cipher` feature
* Move `Loading` to `Features`
