+++ src/Features/Cipher/Autokey/AutokeyStream.tsx
@@ -1,6 +1,6 @@
 import { AutoKeyString } from 'puzzle-lib';
 import React from 'react';
-import KeyedCipherStream from '../../../Common/KeyedCipherStream';
+import KeyedCipherStream from '../KeyedCipherStream';
 
 interface Props {
   prompt: JSX.Element | string;
